docs(api): move binance_best_practices.md to api/ folder for API specific docs EL-1009
Move `binance_best_practices.md` from the root of the `docs/` directory to the `docs/api/` subfolder.

This change reorganizes the `binance_best_practices.md` document, relocating it to the `docs/api/` folder. This is done to better categorize the documentation, as the content of `binance_best_practices.md` is likely focused on best practices specifically related to interacting with the Binance API.

By placing it within the `api/` subfolder, we improve the logical grouping of documentation, making it clearer that these best practices are API-specific and enhancing the discoverability of API-related documentation within the project.

Tuesday 2025-03-25 00:38:07 Vancouver GMT-07:00